添加部分超级用户的功能模块
1.添加查看用户
2.添加修改用户
3.添加删除用户
4.抓蛾子

+++ 项目1/Operateuserdatas.m
@@ -144,11 +144,23 @@
         return tempsta;
     }
     
-    if ([fileop executeUpdate:@"DELETE FROM UserDatas WHERE name = ?",name] == NO )
+    if (name == nil)
     {
-        [fileop close];
-        tempsta = FILEDeleError;
-        return tempsta;
+        if ([fileop executeUpdate:@"DELETE FROM UserDatas"] == NO )
+        {
+            [fileop close];
+            tempsta = FILEDeleError;
+            return tempsta;
+        }
+    }
+    else
+    {
+        if ([fileop executeUpdate:@"DELETE FROM UserDatas WHERE name = ?",name] == NO )
+        {
+            [fileop close];
+            tempsta = FILEDeleError;
+            return tempsta;
+        }
     }
     
     [fileop close];
